+++ examples/c++/x11.cc
@@ -147,7 +147,7 @@ x11::x11( int argc, char **argv ) {
 	{
 	  yy = y[j];
 	  z[i][j] = 3. * (1.-xx)*(1.-xx) * exp(-(xx*xx) - (yy+1.)*(yy+1.)) -
-	    10. * (xx/5. - pow(xx,3.) - pow(yy,5.)) * exp(-xx*xx-yy*yy) -
+	    10. * (xx/5. - pow((double)xx,3.) - pow((double)yy,5.)) * exp(-xx*xx-yy*yy) -
 	    1./3. * exp(-(xx+1)*(xx+1) - (yy*yy));
 	  if(false) { /* Jungfraujoch/Interlaken */
 	    if (z[i][j] < -1.)
